requirements.txt
imageio
keras==2.8.0
Keras-Preprocessing==1.1.2
matplotlib
numpy
opencv-python==4.7.0.72
pandas==1.4.3
Pillow
python-json-logger
PyYAML
scikit-image
scikit-learn
scipy
seaborn
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.2
tensorflow-addons==0.19.0
tensorflow-estimator==2.8.0
tensorflow-io==0.31.0
tensorflow-io-gcs-filesystem==0.31.0
tqdm
tensorflow-addons==0.19.0
pydensecrf